Fix possible infinite loop for crmGetAllCustomers
The crmGetAllCustomers method loads all contacts & all companies from
the Teamleader API in batches of 100 items each. There was just one
small problem that could cause an infinite loop.

The while loops checks if the array is a multiple of 100. If so, the
next page is fetched and so on. But if you had exactly 100 contacts or
companies, or a multiple of 100, every next page returns 0 items. The
result array is always a multiple of 100 since there are no more items
added, and BAM, infinite loop.

This commit adds a check for an empty result set. If so, the while loop
is terminated and the results are returned.